refactor utils, types
