+++ src/unstakeAg/index.ts
@@ -5,15 +5,12 @@ import {
   getAssociatedTokenAddress,
 } from "@solana/spl-token-v2";
 import {
-  AccountInfo,
   Cluster,
   Connection,
-  PublicKey,
   StakeProgram,
   Transaction,
 } from "@solana/web3.js";
 import { Jupiter, JupiterLoadParams, WRAPPED_SOL_MINT } from "@jup-ag/core";
-import { StakeAccount } from "@soceanfi/solana-stake-sdk";
 import { STAKE_ACCOUNT_RENT_EXEMPT_LAMPORTS } from "@soceanfi/stake-pool-sdk";
 import { UnstakeRoute } from "route";
 
@@ -36,6 +33,11 @@ import {
   SOLBLAZE_ADDRESS_MAP,
   UNSTAKE_IT_ADDRESS_MAP,
 } from "@/unstake-ag/unstakeAg/address";
+import type {
+  ComputeRoutesParams,
+  ExchangeParams,
+  ExchangeReturn,
+} from "@/unstake-ag/unstakeAg/types";
 import {
   calcStakeUnstakedAmount,
   chunkedGetMultipleAccountInfos,
@@ -43,11 +45,11 @@ import {
   doTokenProgramAccsExist,
   dummyAccountInfoForProgramOwner,
   genShortestUnusedSeed,
+  outLamports,
+  tryMergeExchangeReturn,
   UNUSABLE_JUP_MARKETS_LABELS,
 } from "@/unstake-ag/unstakeAg/utils";
 import { WithdrawStakePool } from "@/unstake-ag/withdrawStakePools";
-
-export { routeMarketLabels } from "./utils";
 
 export type HybridPool = StakePool & WithdrawStakePool;
 
@@ -513,178 +515,5 @@ export class UnstakeAg {
   }
 }
 
-/**
- *
- * @param param0
- * @returns expected amount of lamports to be received for the given unstake route,
- *          excluding slippage.
- */
-export function outLamports({ stakeAccInput, jup }: UnstakeRoute): bigint {
-  if (!jup) {
-    return stakeAccInput.outAmount;
-  }
-  return BigInt(jup.outAmount.toString());
-}
-
-export function tryMergeExchangeReturn(
-  user: PublicKey,
-  { setupTransaction, unstakeTransaction, cleanupTransaction }: ExchangeReturn,
-): ExchangeReturn {
-  let newSetupTransaction = setupTransaction;
-  let newUnstakeTransaction = unstakeTransaction;
-  let newCleanupTransaction = cleanupTransaction;
-
-  if (setupTransaction) {
-    const mergeSetup = tryMerge2Txs(user, setupTransaction, unstakeTransaction);
-    if (mergeSetup) {
-      newSetupTransaction = undefined;
-      newUnstakeTransaction = mergeSetup;
-    }
-  }
-
-  if (cleanupTransaction) {
-    const mergeCleanup = tryMerge2Txs(
-      user,
-      newUnstakeTransaction,
-      cleanupTransaction,
-    );
-    if (mergeCleanup) {
-      newCleanupTransaction = undefined;
-      newUnstakeTransaction = mergeCleanup;
-    }
-  }
-  return {
-    setupTransaction: newSetupTransaction,
-    unstakeTransaction: newUnstakeTransaction,
-    cleanupTransaction: newCleanupTransaction,
-  };
-}
-
-/**
- * TODO: Check that additional signatures required are accounted for
- * i.e. actual tx size is `tx.serialize().length`
- * and not `tx.serialize().length + 64 * additional required signatures`
- * @param feePayer
- * @param firstTx
- * @param secondTx
- * @returns a new transaction with the 2 transaction's instructions merged if possible, null otherwise
- */
-function tryMerge2Txs(
-  feePayer: PublicKey,
-  firstTx: Transaction,
-  secondTx: Transaction,
-): Transaction | null {
-  const MOCK_BLOCKHASH = "41xkyTsFaxnPvjv3eJMdjGfmQj3osuTLmqC3P13stSw3";
-  const SERIALIZE_CONFIG = {
-    requireAllSignatures: false,
-    verifyAllSignatures: false,
-  };
-  const merged = new Transaction();
-  merged.add(...firstTx.instructions);
-  merged.add(...secondTx.instructions);
-  merged.feePayer = feePayer;
-  merged.recentBlockhash = MOCK_BLOCKHASH;
-  try {
-    merged.serialize(SERIALIZE_CONFIG);
-  } catch (e) {
-    const msg = (e as Error).message;
-    if (msg && msg.includes("Transaction too large")) {
-      return null;
-    }
-    // uncaught
-    throw e;
-  }
-  merged.feePayer = undefined;
-  merged.recentBlockhash = undefined;
-  return merged;
-}
-
-export interface ComputeRoutesParams {
-  /**
-   * The stake account to be unstaked
-   */
-  stakeAccount: AccountInfo<StakeAccount>;
-
-  /**
-   * The amount in lamports to be unstaked.
-   * Should be <= stakeAccount.lamports.
-   * If < stakeAccount.lamports, a stake split instruction will be
-   * added to the setup instructions
-   */
-  amountLamports: bigint;
-
-  /**
-   * In basis point (0 - 10_000)
-   */
-  slippageBps: number;
-
-  /**
-   * Same as `jupiter.computeRoutes()` 's `forceFetch`:
-   * If true, refetches all jup accounts and stake pool accounts
-   *
-   * Default to false
-   */
-  forceFetch?: boolean;
-
-  /**
-   * Optional additional fee to charge on jup swaps,
-   * passed as `feeBps` to `jupiter.computeRoutes()`
-   *
-   * Defaults to undefined
-   */
-  jupFeeBps?: number;
-
-  /**
-   * Silently ignore routes where errors were thrown
-   * during computation such as failing to fetch
-   * required accounts.
-   *
-   * Defaults to true
-   */
-  shouldIgnoreRouteErrors?: boolean;
-}
-
-export interface ExchangeParams {
-  /**
-   * A route returned by `computeRoutes()`
-   */
-  route: UnstakeRoute;
-
-  /**
-   * Fetched on-chain data of the stake account
-   * to unstake
-   */
-  stakeAccount: AccountInfo<StakeAccount>;
-
-  /**
-   * Pubkey of the stake account to unstake
-   */
-  stakeAccountPubkey: PublicKey;
-
-  /**
-   * Withdraw authority of the stake account to unstake
-   */
-  user: PublicKey;
-
-  /**
-   * Wrapped SOL account to receive optional additional fee on
-   * jup swaps when `jupFeeBps` is set on `computeRoutes()`
-   */
-  feeAccounts?: FeeAccounts;
-}
-
-/**
- * Map of token mint to token accounts to receive referral fees
- * If token is So11111111111111111111111111111111111111112,
- * the value should be a wrapped SOL token account, not a system account.
- * SyncNative is not guaranteed to be called after transferring SOL referral fees
- */
-export type FeeAccounts = {
-  [token in string]?: PublicKey;
-};
-
-export interface ExchangeReturn {
-  setupTransaction?: Transaction;
-  unstakeTransaction: Transaction;
-  cleanupTransaction?: Transaction;
-}
+export * from "./types";
+export * from "./utils";
